THRIFT-1070 C++ compiler and runtime have 32/64bit problems Patch: Rich Salz
git-svn-id: https://svn.apache.org/repos/asf/thrift/trunk@1075121 13f79535-47bb-0310-9956-ffa450edef68

+++ lib/cpp/src/transport/TBufferTransports.h
@@ -108,7 +108,7 @@ class TBufferBase : public TVirtualTransport<TBufferBase> {
     if (TDB_LIKELY(static_cast<ptrdiff_t>(*len) <= rBound_ - rBase_)) {
       // With strict aliasing, writing to len shouldn't force us to
       // refetch rBase_ from memory.  TODO(dreiss): Verify this.
-      *len = rBound_ - rBase_;
+      *len = static_cast<uint32_t>(rBound_ - rBase_);
       return rBase_;
     }
     return borrowSlow(buf, len);
@@ -568,7 +568,7 @@ class TMemoryBuffer : public TVirtualTransport<TMemoryBuffer, TBufferBase> {
   // TODO(dreiss): Make bufPtr const.
   void getBuffer(uint8_t** bufPtr, uint32_t* sz) {
     *bufPtr = rBase_;
-    *sz = wBase_ - rBase_;
+    *sz = static_cast<uint32_t>(wBase_ - rBase_);
   }
 
   std::string getBufferAsString() {
@@ -656,11 +656,11 @@ class TMemoryBuffer : public TVirtualTransport<TMemoryBuffer, TBufferBase> {
 
   uint32_t available_read() const {
     // Remember, wBase_ is the real rBound_.
-    return wBase_ - rBase_;
+    return static_cast<uint32_t>(wBase_ - rBase_);
   }
 
   uint32_t available_write() const {
-    return wBound_ - wBase_;
+    return static_cast<uint32_t>(wBound_ - wBase_);
   }
 
   // Returns a pointer to where the client can write data to append to
